use DI for SettingsWindow
